Nguyen manage tour

+++ backend/src/controllers/news.controller.js
@@ -4,11 +4,22 @@ const { newsService } = require('../services')
 const { newsValidation } = require('../validations')
 const fs = require('fs')
 
+
+/* create new news */
 const createNews = catchAsync(async(req, res) => {
-    const newsBody = Object.assign(req.body, { imageUrl: req.file.path })
+    const image = req.file ? { imageUrl: req.file.path } : {}
+    const newsBody = Object.assign(req.body, image)
     const validation = await newsValidation.validate(newsBody)
     if (validation.error) {
         const errorMessage = validation.error.details[0].message
+        if (req.file) {
+            fs.unlink(`${req.file.path}`, (err) => {
+                if (err) {
+                    console.error(err)
+                    res.status(httpStatus.BAD_REQUEST).send({ message: err })
+                }
+            })
+        }
         return res.status(httpStatus.BAD_REQUEST).send({
             message: errorMessage
         })
@@ -17,6 +28,7 @@ const createNews = catchAsync(async(req, res) => {
     res.status(httpStatus.CREATED).send(news)
 })
 
+/* get all news */
 const getNewsByPage = catchAsync(async(req, res) => {
     const perPage = 6;
     let page = parseInt(req.query.page) || 1;
@@ -27,8 +39,9 @@ const getNewsByPage = catchAsync(async(req, res) => {
     } else res.status(200).json({ news, totalNews })
 })
 
+
+/* get news detail by params id */
 const getNewsById = catchAsync(async(req, res) => {
-    
     const newsSingle = await newsService.getNewsById(req.params.id)
 
     if (!newsSingle) {
@@ -36,20 +49,34 @@ const getNewsById = catchAsync(async(req, res) => {
     } else res.send(newsSingle);
 })
 
+/* update news detail by params id*/
 const updateNewsById = catchAsync(async(req, res) => {
-    const newsSingle = await newsService.getNewsById(req.params.id)
-    const path = newsSingle.imageUrl.slice(14)
-    fs.unlink(`./public/uploads/${path}`, (err) => {
-        if (err) {
-            console.error(err)
-            res.status(httpStatus.BAD_REQUEST).send({ error: err })
+    if (req.file) {
+        const newsSingle = await newsService.getNewsById(req.params.id)
+        if (tourData.imageUrl) {
+            const path = newsSingle.imageUrl.slice(14)
+            fs.unlink(`./public/uploads/${path}`, (err) => {
+                if (err) {
+                    console.error(err)
+                    res.status(httpStatus.BAD_REQUEST).send({ message: err })
+                }
+            })
         }
-    })
+    }
 
-    const newsBody = Object.assign(req.body, { imageUrl: req.file.path })
+    const image = req.file ? { imageUrl: req.file.path } : {}
+    const newsBody = Object.assign(req.body, image)
     const validation = await newsValidation.validate(newsBody)
     if (validation.error) {
         const errorMessage = validation.error.details[0].message
+        if (req.file) {
+            fs.unlink(`${req.file.path}`, (err) => {
+                if (err) {
+                    console.error(err)
+                    res.status(httpStatus.BAD_REQUEST).send({ message: err })
+                }
+            })
+        }
         return res.status(httpStatus.BAD_REQUEST).send({
             message: errorMessage
         })
@@ -62,9 +89,23 @@ const updateNewsById = catchAsync(async(req, res) => {
     res.status(200).send(news)
 })
 
+
+/* delete news by params id */
 const deleteNewsById = catchAsync(async(req, res) => {
+    const newsData = await newsService.getNewsById(req.params.id)
+    if (!newsData) {
+        res.status(httpStatus.NOT_FOUND).send("News not found")
+    }
+    if (newsData.imageUrl) {
+        fs.unlink(`${newsData.imageUrl}`, (err) => {
+            if (err) {
+                console.error(err)
+                res.status(httpStatus.BAD_REQUEST).send({ message: err })
+            }
+        })
+    }
     await newsService.deleteNewsById(req.params.id)
-    res.status(httpStatus.NO_CONTENT).send("News has been deleted")
+    res.status(httpStatus.NO_CONTENT).send()
 })
